Updating development status
I've added unit testing now, and things are progressing smoothly. I hope
to get a working prototype soon that will at least work in a hard coded
state. Once everything works in that hard coded state, it will be time
to code this thing out of hack code mode and into production code. With
the correct code in place, I will feel much more confident in the status
of this project.