Created command parser for process.cpp
I have created the command and argument parsing for process.cpp. We
should now be able to pass a string into the constructor and get a valid
char array containing all of the arguments to send to the exec c
function. Hopefully the windows/mac APIs will accept the same style of
input, but we will see once we get to that.