Added directory listing utility
I needed a way to list directories and be able to filter files/directories to
perform the directory watching that I need. The specific use case that requires
this function is that I need to be able to recurse through directories and add
directory watchers to each directory via inotify.