Organized source files
I've separated the source files from the tests files and other miscellaneous
documentation files. I did this so it would be easier to keep track of where
all of the different source files are located. It is quite confusing when they
are all mixed together.

I have also taken the liberty to start refactoring the configuration file
parser to use a tokenizer approach, where each structure/class handles it's own
section by itself.