Finally got renzoku 【れんぞく】 to work
After spending most of the day trying to hack away at this code to get it to
function in a most basic state, I have finally done it. :) renzoku is finally
working! :D I'm so excited! From here on out, the priorities have shifted from
"get it to work" to "get it to work well"! Yipee! My very first actual pet
project that hasn't fallen by the wayside.

Signed-off-by: Colton Wolkins (Ogre) <frostyfrog2@gmail.com>