Tweaks for Ubuntu
In an attempt to make sure that Renzoku could run on multiple different
distributions, I have made a Ubuntu VM and tweaked Renzoku until it would run
on Ubuntu, just like how it runs on Archlinux. There have been some changes to
the way that g++ compiles code that allowed Renzoku to run and compile on
Archlinux, though it failed to do so on Ubuntu. These tweaks should fix that.

Signed-off-by: Colton Wolkins (Ogre) <frostyfrog2@gmail.com>