Switched from yaml config to libconfig
Switched renzoku from using yaml based configs to libconfig instead. libconfig
is much more in line with the type of config that I was hoping to have renzoku
use in the first place. Now that I have finally decided on an official config
format, I can finally start actually recommending it's use to others since
renzoku should be able to build on other OSs. It used to be that only Linux
Distros with the latest and greatest packages could do so.

This fixes #2 and gets a closer to completing the tasks needed for #1.

Signed-off-by: Colton Wolkins (Ogre) <frostyfrog2@gmail.com>