Added first pass of config parser
Detailed Reasoning
==================

Fed Up
------

I'm fed up with the configuration parsers out there. The versions that are out
there that I've used in this project so far (libconfig++ and libyaml-cpp) have
changed their API in some way or another that made it impossible to compile on
older systems, without shipping the libraries along with the package or
statically linking them into renzoku.

Birth of a Parser
-----------------

I have started to write my own parser so that I A) achieve better
cross-platformify renzoku and B) finally learn how to write a parser from the
ground up. The code within this commit is still in-development and will be
until it can successfully parse the config file outlined in GH-3.

Problems with the previous config format
----------------------------------------

Quite frankly, the main problem with the previous config file format was that I
was unable to step away from the project and think of other use cases. The only
main use cases that I had were my own. Speaking about Renzoku at PLUG really
helped me understand how renzoku could be used by other people. GH-3 came out
of PLUG and has become more of a guide for where I'm heading next with my code.
:D