Added Official Parser
This is the first pass at my official parser. It isn't complete, it's not near
complete, but it works well so far. I'm creating and pushing this commit as a
checkpoint since I'm currently at 801Labs and I want to make sure that I have a
backup somewhere. :) Never know what hacktivists might do ;p